6: Prevent race conditions in model/Quote.java
Task-Url: https://github.com/ZIIIM/babybash/issues/issues/6

+++ src/main/java/ziiim/babybash/repository/QuoteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import ziiim.babybash.model.Quote;
 
 @Repository
-public interface QuoteRepository extends JpaRepository<Quote, Integer>
+public interface QuoteRepository extends JpaRepository<Quote, Long>
 {
 	public List<Quote> findFirst10ByPublishedIsTrueOrderByCreationDateDesc();
 }
